+++ dev/main-react16/src/global.jsx
@@ -3,42 +3,46 @@ import microApp, { unmountApp, unmountAllApps } from '@micro-zoe/micro-app'
 import config from './config'
 
 const prefetchConfig = [
-  {
-    name: 'vite',
-    url: `${config.vite}micro-app/vite`,
-    esmodule: true,
-    // inline: true,
-    // 'disable-sandbox': true,
-  },
+  // {
+  //   name: 'vite',
+  //   url: `${config.vite}micro-app/vite`,
+  //   esmodule: true,
+  //   // inline: true,
+  //   // 'disable-sandbox': true,
+  // },
   {
     name: 'vue2',
     url: `${config.vue2}micro-app/vue2`,
     // 'disable-scopecss': true,
+    preRender: {
+      'default-page': '/micro-app/vue2/#/page2',
+      'disable-patch-request': false,
+    }
   },
-  {
-    name: 'react16',
-    url: `${config.react16}micro-app/react16`,
-  },
-  {
-    name: 'react17',
-    url: `${config.react17}micro-app/react17`,
-  },
-  {
-    name: 'vue3',
-    url: `${config.vue3}micro-app/vue3`,
-  },
-  {
-    name: 'angular11',
-    url: `${config.angular11}micro-app/angular11`,
-  },
-  {
-    name: 'angular14',
-    url: `${config.angular14}micro-app/angular14`,
-    esmodule: true,
-  },
+  // {
+  //   name: 'react16',
+  //   url: `${config.react16}micro-app/react16`,
+  // },
+  // {
+  //   name: 'react17',
+  //   url: `${config.react17}micro-app/react17`,
+  // },
+  // {
+  //   name: 'vue3',
+  //   url: `${config.vue3}micro-app/vue3`,
+  // },
+  // {
+  //   name: 'angular11',
+  //   url: `${config.angular11}micro-app/angular11`,
+  // },
+  // {
+  //   name: 'angular14',
+  //   url: `${config.angular14}micro-app/angular14`,
+  //   esmodule: true,
+  // },
 ]
 
-// microApp.preFetch(prefetchConfig)
+microApp.preFetch(prefetchConfig)
 
 microApp.start({
   // shadowDOM: true,
@@ -185,7 +189,7 @@ window.addEventListener('popstate', (e) => {
   // const a = document.createElement('div')
   //   a.innerHTML = '55555555'
   //   document.body.appendChild(a)
-  console.log('popstate', e, window.location.href)
+  console.log('基座 popstate', e, window.location.href)
   // history.replaceState(history.state, '', location.href)
 })
 
@@ -193,5 +197,5 @@ window.addEventListener('hashchange', (e) => {
   // const a = document.createElement('div')
   //   a.innerHTML = '666666666'
   //   document.body.appendChild(a)
-  console.log('hashchange', e, e.newURL, e.oldURL)
+  console.log('基座 hashchange', e, e.newURL, e.oldURL)
 })