+++ dev/main-react16/src/pages/vue2/vue2.js
@@ -86,7 +86,7 @@ function Vue2 () {
         // inline
         // disable-scopecss
         // disableSandbox
-        clear-data
+        // clear-data
       >
       </micro-app>
     </div>