+++ dev/main-react16/src/pages/vue3/vue3.js
@@ -28,9 +28,9 @@ function Vue3 () {
     <div>
       <Col span={7} className='btn-con'>
         <Button type="primary" onClick={changeData.bind(this, {type: '新数据'})}>向子应用发送数据</Button>
-        <Button type="primary" onClick={jumpToHome}>基座控制子应用跳转home</Button>
-        <Button type="primary" onClick={jumpToElementPlus}>基座控制子应用跳转element-plus</Button>
-        <Button type="primary" onClick={jumpToAntDesignVue}>基座控制子应用跳转ant-design-vue</Button>
+        <Button type="primary" onClick={jumpToHome}>控制子应用跳转home</Button>
+        <Button type="primary" onClick={jumpToElementPlus}>控制子应用跳转element-plus</Button>
+        <Button type="primary" onClick={jumpToAntDesignVue}>控制子应用跳转ant-design-vue</Button>
       </Col>
       {
         showLoading && <Spin indicator={antIcon} />
@@ -43,7 +43,7 @@ function Vue3 () {
         onBeforeshow={() => hideLoading(false)}
         // baseRoute='/micro-app/demo/vue3'
         // disableScopecss
-        // keep-alive
+        keep-alive
         keep-router-state
         // inline
         // destroy